# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CommonGrid API",
    "version": "1.0.0",
...
    "contact": {
      "name": "CommonGrid Support",
      "email": "hello@texturehq.com",
      "url": "https://texturehq.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/TextureHQ/commongrid/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://commongrid.info/api/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from the Developer Dashboard. Prefix with `cg_`. Only Bearer is accepted; `X-API-Key` and other Authorization schemes are not supported."
      }
    },
    "parameters": {
...
      "slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "Unique URL-safe identifier for the resource",
        "schema": {
          "type": "string"
        }
      },
      "at": {
        "name": "at",
        "in": "query",
        "description": "Point-in-time read. Reconstructs the entity from version history at the latest version with changed_at ≤ this instant. Accepts YYYY-MM-DD (end of that UTC day) or an ISO-8601 timestamp. Response includes `_as_of: { requested, versionNumber, changedAt }`. Returns 404 when no version exists at or before the requested time.",
        "schema": {
          "type": "string"
        },
        "example": "2025-12-01"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          }
        }
      },
      "RateLimitedError": {
        "type": "object",
        "description": "Returned when either the hourly or burst rate-limit window is exceeded. Header values describe the window that tripped.",
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "currentTier",
              "currentLimit",
              "retryAfter"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "RATE_LIMITED"
                ],
                "example": "RATE_LIMITED"
              },
              "message": {
                "type": "string",
                "example": "Too many requests. Please slow down."
              },
              "request_id": {
                "type": "string",
                "example": "req_abc123"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "currentTier": {
                "type": "string",
                "enum": [
                  "anonymous",
                  "registered",
                  "bulk",
                  "write"
                ],
                "example": "registered"
              },
              "currentLimit": {
                "type": "integer",
                "description": "Budget of the window that was enforced (hourly or burst)",
                "example": 100
              },
              "retryAfter": {
                "type": "integer",
                "description": "Seconds until X-RateLimit-Reset for the window that tripped",
                "example": 12
              },
              "docs": {
                "type": "string",
                "format": "uri"
              },
              "upgrade": {
                "type": "string",
                "description": "Present on anonymous 429s — points callers at free API key registration"
              }
            }
          }
        }
      },
...
      "Rto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable internal identifier"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe unique identifier used in API paths"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "shortName": {
            "type": "string",
            "description": "Short name / abbreviation"
          },
          "logo": {
            "type": "string",
            "nullable": true,
            "description": "Logo URL"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "description": "Primary website"
          },
          "states": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of two-letter US state codes covered"
          },
          "regionId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the parent region"
          },
          "source": {
            "type": "string",
            "nullable": true,
            "description": "Upstream data source identifier (e.g. EIA, HIFLD, OSM)"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Canonical source URL (filing PDF, dataset download, etc.)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was first ingested into CommonGrid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was last modified"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Soft-delete timestamp (null for active records)"
          },
          "version": {
            "type": "integer",
            "description": "Monotonic version counter incremented on every edit"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "shortName",
          "states",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "BalancingAuthority": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable internal identifier"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe unique identifier used in API paths"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "shortName": {
            "type": "string",
            "description": "Short name / abbreviation"
          },
          "logo": {
            "type": "string",
            "nullable": true,
            "description": "Logo URL"
          },
          "eiaCode": {
            "type": "string",
            "nullable": true,
            "description": "EIA balancing authority code"
          },
          "eiaId": {
            "type": "string",
            "nullable": true,
            "description": "EIA identifier"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "description": "Primary website"
          },
          "states": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of two-letter US state codes covered"
          },
          "isoId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the parent ISO, if any"
          },
          "regionId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the region this BA sits inside"
          },
          "source": {
            "type": "string",
            "nullable": true,
            "description": "Upstream data source identifier (e.g. EIA, HIFLD, OSM)"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Canonical source URL (filing PDF, dataset download, etc.)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was first ingested into CommonGrid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was last modified"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Soft-delete timestamp (null for active records)"
          },
          "version": {
            "type": "integer",
            "description": "Monotonic version counter incremented on every edit"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "shortName",
          "states",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "Region": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable internal identifier"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe unique identifier used in API paths"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "type": {
            "type": "string",
            "description": "Region type (state, county, iso, rto, ba, utility-service-territory, cca, etc.)"
          },
          "eiaId": {
            "type": "string",
            "nullable": true,
            "description": "EIA identifier, if applicable"
          },
          "state": {
            "type": "string",
            "nullable": true,
            "description": "Two-letter US state code, if applicable"
          },
          "customers": {
            "type": "integer",
            "nullable": true,
            "description": "Customer count associated with this region, if reported"
          },
          "source": {
            "type": "string",
            "nullable": true,
            "description": "Upstream data source identifier (e.g. EIA, HIFLD, OSM)"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Canonical source URL (filing PDF, dataset download, etc.)"
          },
          "sourceDate": {
            "type": "string",
            "nullable": true,
            "description": "Source filing date (YYYY-MM-DD)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was first ingested into CommonGrid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was last modified"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Soft-delete timestamp (null for active records)"
          },
          "version": {
            "type": "integer",
            "description": "Monotonic version counter incremented on every edit"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "type",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
    }
  },
  "paths": {
...
    "/rtos/{slug}": {
      "get": {
        "operationId": "getRto",
        "summary": "Get RTO by slug",
        "tags": [
          "Grid Operators"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          },
          {
            "$ref": "#/components/parameters/at"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Rto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Hourly or burst rate-limit window exceeded. `Retry-After` and `X-RateLimit-*` describe the window that tripped.",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying. Matches X-RateLimit-Reset for the window that returned 429 (burst or hourly)."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitedError"
                }
              }
            }
          }
        }
      }
    },
...
    "/balancing-authorities/{slug}": {
      "get": {
        "operationId": "getBalancingAuthority",
        "summary": "Get balancing authority by slug",
        "tags": [
          "Grid Operators"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          },
          {
            "$ref": "#/components/parameters/at"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BalancingAuthority"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Hourly or burst rate-limit window exceeded. `Retry-After` and `X-RateLimit-*` describe the window that tripped.",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying. Matches X-RateLimit-Reset for the window that returned 429 (burst or hourly)."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitedError"
                }
              }
            }
          }
        }
      }
    },
...
    "/regions/{slug}": {
      "get": {
        "operationId": "getRegion",
        "summary": "Get region by slug",
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          },
          {
            "$ref": "#/components/parameters/at"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Region"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Hourly or burst rate-limit window exceeded. `Retry-After` and `X-RateLimit-*` describe the window that tripped.",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying. Matches X-RateLimit-Reset for the window that returned 429 (burst or hourly)."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitedError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}